+++ src/main/java/com/softserve/academy/service/UserService.java
@@ -9,9 +9,9 @@ public interface UserService {
 
     User getUserById(Integer id) throws IllegalArgumentException;
 
-    double getUserStatisticAverageAge();
+    int getUserStatisticAverageAge();
 
-    double getUserAverageTimeOfUsingLibrary();
+    int getUserAverageTimeOfUsingLibrary();
 
     User getRegisteredUser(String username, String password) throws IllegalArgumentException;
 }
